+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "didibudget-backend",
-  "version": "0.5.12",
+  "version": "0.5.13",
   "description": "didibudget-backend",
   "main": "src/server.js",
   "engines": {
