+++ src/gql/resolvers/expenses.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { logger } = require('../../utils/logger');
+const { logger } = require('../../helpers/logger');
 
 const { Expenses } = require('../../data/models/index');
 const { authValidations } = require('../auth/authValidations');
